Factor some duplicated code
