+++ js/source/raster_tile_source.js
@@ -2,46 +2,16 @@
 
 var util = require('../util/util');
 var ajax = require('../util/ajax');
-var browser = require('../util/browser');
-var normalizeURL = require('../util/mapbox').normalizeSourceURL;
 var Evented = require('../util/evented');
-var TilePyramid = require('./tile_pyramid');
 var TileCoord = require('./tile_coord');
+var Source = require('./source');
 
 module.exports = RasterTileSource;
 
 function RasterTileSource(options) {
     util.extend(this, util.pick(options, 'url', 'tileSize'));
 
-    var loaded = (err, tileJSON) => {
-        if (err) {
-            this.fire('error', {error: err});
-            return;
-        }
-
-        util.extend(this, util.pick(tileJSON,
-            'tiles', 'minzoom', 'maxzoom', 'attribution'));
-
-        this._pyramid = new TilePyramid({
-            tileSize: this.tileSize,
-            cacheSize: 20,
-            minzoom: this.minzoom,
-            maxzoom: this.maxzoom,
-            load: this._loadTile.bind(this),
-            abort: this._abortTile.bind(this),
-            unload: this._unloadTile.bind(this),
-            add: this._addTile.bind(this),
-            remove: this._removeTile.bind(this)
-        });
-
-        this.fire('load');
-    };
-
-    if (this.url) {
-        ajax.getJSON(normalizeURL(this.url), loaded);
-    } else {
-        browser.frame(loaded.bind(this, null, options));
-    }
+    Source._loadTileJSON.call(this, options);
 }
 
 RasterTileSource.prototype = util.inherit(Evented, {
@@ -64,13 +34,7 @@ RasterTileSource.prototype = util.inherit(Evented, {
         }
     },
 
-    render(layers, painter) {
-        if (this._pyramid) {
-            this._pyramid.renderedIDs().forEach((id) => {
-                painter.drawTile(id, this._pyramid.getTile(id), layers);
-            });
-        }
-    },
+    render: Source._renderTiles,
 
     _loadTile(tile) {
         ajax.getImage(TileCoord.url(tile.id, this.tiles), (err, img) => {
