+++ js/source/vector_tile_source.js
@@ -1,12 +1,9 @@
 'use strict';
 
 var util = require('../util/util');
-var ajax = require('../util/ajax');
-var browser = require('../util/browser');
-var normalizeURL = require('../util/mapbox').normalizeSourceURL;
 var Evented = require('../util/evented');
-var TilePyramid = require('./tile_pyramid');
 var TileCoord = require('./tile_coord');
+var Source = require('./source');
 
 module.exports = VectorTileSource;
 
@@ -17,35 +14,7 @@ function VectorTileSource(options) {
         throw new Error('vector tile sources must have a tileSize of 512');
     }
 
-    var loaded = (err, tileJSON) => {
-        if (err) {
-            this.fire('error', {error: err});
-            return;
-        }
-
-        util.extend(this, util.pick(tileJSON,
-            'tiles', 'minzoom', 'maxzoom', 'attribution'));
-
-        this._pyramid = new TilePyramid({
-            tileSize: 512,
-            minzoom: this.minzoom,
-            maxzoom: this.maxzoom,
-            cacheSize: 20,
-            load: this._loadTile.bind(this),
-            abort: this._abortTile.bind(this),
-            unload: this._unloadTile.bind(this),
-            add: this._addTile.bind(this),
-            remove: this._removeTile.bind(this)
-        });
-
-        this.fire('load');
-    };
-
-    if (this.url) {
-        ajax.getJSON(normalizeURL(this.url), loaded);
-    } else {
-        browser.frame(loaded.bind(this, null, options));
-    }
+    Source._loadTileJSON.call(this, options);
 }
 
 VectorTileSource.prototype = util.inherit(Evented, {
@@ -68,13 +37,8 @@ VectorTileSource.prototype = util.inherit(Evented, {
         }
     },
 
-    render(layers, painter) {
-        if (this._pyramid) {
-            this._pyramid.renderedIDs().forEach((id) => {
-                painter.drawTile(id, this._pyramid.getTile(id), layers);
-            });
-        }
-    },
+    render: Source._renderTiles,
+    featuresAt: Source._vectorFeaturesAt,
 
     _loadTile(tile) {
         var params = {
@@ -117,23 +81,5 @@ VectorTileSource.prototype = util.inherit(Evented, {
         tile.unloadVectorData(this.map.painter);
         this.glyphAtlas.removeGlyphs(tile.uid);
         this.dispatcher.send('remove tile', { id: tile.uid, source: this.id }, null, tile.workerID);
-    },
-
-    featuresAt(point, params, callback) {
-        if (!this._pyramid)
-            return callback(null, []);
-
-        var result = this._pyramid.tileAt(point);
-        if (!result)
-            return callback(null, []);
-
-        this.dispatcher.send('query features', {
-            id: result.tile.uid,
-            x: result.x,
-            y: result.y,
-            scale: result.scale,
-            source: this.id,
-            params: params
-        }, callback, result.tile.workerID);
     }
 });
